+++ server/src/dtsTypes/standardTypes.ts
@@ -23,6 +23,7 @@ import { getReservedMemoryNodeType } from "./standardTypes/nodeTypes/reserved-me
 import { getChosenNodeType } from "./standardTypes/nodeTypes/chosen/node";
 import { getCpusNodeType } from "./standardTypes/nodeTypes/cpus/node";
 import { getCpuNodeType } from "./standardTypes/nodeTypes/cpus/cpu/node";
+import { getReservedMemoryChildNodeType } from "./standardTypes/nodeTypes/reserved-memory/child/node";
 
 export function getStandardType(node?: Node) {
   switch (node?.name) {
@@ -40,6 +41,10 @@ export function getStandardType(node?: Node) {
       return getCpusNodeType();
     case "cpu":
       return getCpuNodeType();
+  }
+
+  if (node?.parent?.name === "reserved-memory") {
+    return getReservedMemoryChildNodeType();
   }
 
   return getStandardDefaultType();