+++ server/src/dtsTypes/standardTypes/nodeTypes/reserved-memory/child/noMap.ts
@@ -14,29 +14,19 @@
  * limitations under the License.
  */
 
-import { BindingPropertyType } from "../../../../types/index";
-import { PropertyNodeType } from "../../../types";
-import { generateOrTypeObj } from "../../helpers";
+import { BindingPropertyType } from "../../../../../types/index";
+import { PropertyNodeType } from "../../../../types";
+import { generateOrTypeObj } from "../../../helpers";
 
 export default () => {
   const prop = new PropertyNodeType(
-    "chassis-type",
-    generateOrTypeObj(BindingPropertyType.STRING),
-    "optional",
-    undefined,
-    [
-      "desktop",
-      "laptop",
-      "convertible",
-      "server",
-      "tablet",
-      "handset",
-      "watch",
-      "embedded",
-    ]
+    "no-map",
+    generateOrTypeObj(BindingPropertyType.EMPTY),
+    "optional"
   );
   prop.description = [
-    `Specifies a string that identifies the form-factor of the system.`,
+    `Indicates the operating system must not create a virtual mapping of the region as part of its standard mapping of system memory, nor permit speculative access to it under
+any circumstances other than under the control of the device driver using the region.`,
   ];
 
   return prop;