+++ server/src/dtsTypes/standardTypes/nodeTypes/reserved-memory/child/node.ts
@@ -14,30 +14,28 @@
  * limitations under the License.
  */
 
-import { BindingPropertyType } from "../../../../types/index";
-import { PropertyNodeType } from "../../../types";
-import { generateOrTypeObj } from "../../helpers";
+import { getStandardDefaultType } from "../../../../standardDefaultType";
+import alignment from "./alignment";
+import allocRanges from "./allocRanges";
+import noMap from "./noMap";
+import reusable from "./reusable";
+import size from "./size";
 
-export default () => {
-  const prop = new PropertyNodeType(
-    "chassis-type",
-    generateOrTypeObj(BindingPropertyType.STRING),
-    "optional",
-    undefined,
-    [
-      "desktop",
-      "laptop",
-      "convertible",
-      "server",
-      "tablet",
-      "handset",
-      "watch",
-      "embedded",
-    ]
-  );
-  prop.description = [
-    `Specifies a string that identifies the form-factor of the system.`,
-  ];
+export function getReservedMemoryChildNodeType() {
+  const nodeType = getStandardDefaultType();
 
-  return prop;
-};
+  const regProp = nodeType.properties.find((p) => p.name === "reg");
+  regProp!.required = () => {
+    return "optional";
+  };
+
+  nodeType.addProperty([
+    size(),
+    alignment(),
+    allocRanges(),
+    noMap(),
+    reusable(),
+  ]);
+
+  return nodeType;
+}