+++ server/src/dtsTypes/standardTypes/nodeTypes/reserved-memory/child/reusable.ts
@@ -14,29 +14,18 @@
  * limitations under the License.
  */
 
-import { BindingPropertyType } from "../../../../types/index";
-import { PropertyNodeType } from "../../../types";
-import { generateOrTypeObj } from "../../helpers";
+import { BindingPropertyType } from "../../../../../types/index";
+import { PropertyNodeType } from "../../../../types";
+import { generateOrTypeObj } from "../../../helpers";
 
 export default () => {
   const prop = new PropertyNodeType(
-    "chassis-type",
-    generateOrTypeObj(BindingPropertyType.STRING),
-    "optional",
-    undefined,
-    [
-      "desktop",
-      "laptop",
-      "convertible",
-      "server",
-      "tablet",
-      "handset",
-      "watch",
-      "embedded",
-    ]
+    "reusable",
+    generateOrTypeObj(BindingPropertyType.EMPTY),
+    "optional"
   );
   prop.description = [
-    `Specifies a string that identifies the form-factor of the system.`,
+    `The operating system can use the memory in this region with the limitation that the device driver(s) owning the region need to be able to reclaim it back. Typically that means that the operating system can use that region to store volatile or cached data that can be otherwise regenerated or migrated elsewhere.`,
   ];
 
   return prop;